refactor(util): move `ConcurrencyLimiter` to its own export
