+++ packages/@sanity/util/src/client/concurrency-limiter/index.ts
@@ -1,2 +1,1 @@
-export * from './ConcurrencyLimiter'
 export * from './createClientConcurrencyLimiter'
