+++ packages/@sanity/util/turbo.json
@@ -7,6 +7,7 @@
         "lib/**",
         "index.js",
         "client.js",
+        "ConcurrencyLimiter.js",
         "content.js",
         "createSafeJsonParser.js",
         "fs.js",
