+++ packages/sanity/src/core/validation/util/createBatchedGetDocumentExists.ts
@@ -13,7 +13,7 @@ import {
 } from 'rxjs'
 
 import {type AvailabilityResponse} from '../../preview'
-import {ConcurrencyLimiter} from '@sanity/util/client'
+import {ConcurrencyLimiter} from '@sanity/util/ConcurrencyLimiter'
 
 /**
  * The amount of time reserved for waiting for new IDs.
