+++ packages/sanity/src/core/validation/validateDocument.ts
@@ -13,7 +13,8 @@ import {concat, defer, from, lastValueFrom, merge, Observable, of} from 'rxjs'
 import {catchError, map, mergeAll, mergeMap, switchMap, toArray} from 'rxjs/operators'
 
 import {type SourceClientOptions, type Workspace} from '../config'
-import {ConcurrencyLimiter, createClientConcurrencyLimiter} from '@sanity/util/client'
+import {createClientConcurrencyLimiter} from '@sanity/util/client'
+import {ConcurrencyLimiter} from '@sanity/util/ConcurrencyLimiter'
 import {getFallbackLocaleSource} from '../i18n/fallback'
 import {type ValidationContext} from './types'
 import {createBatchedGetDocumentExists} from './util/createBatchedGetDocumentExists'
